+++ app/scripts/components/extended/add-box.js
@@ -7,7 +7,7 @@ angular.module("korpApp").component("addBox", {
         <div class="mt-10 mr-14 inline-block" ng-mouseleave="$ctrl.reset()">
             <button
                 ng-class="{'fade-out': $ctrl.showStuffSelectButtons}"
-                class="btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover_bg-gray-200"
+                class="btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover:bg-gray-200"
                 ng-click="$ctrl.addTokenLocal()"
             >
                 <i class="fa-solid fa-lg fa-plus-circle text-blue-600 mr-1"></i>
@@ -16,7 +16,7 @@ angular.module("korpApp").component("addBox", {
             <div class="mt-2">
                 <button
                     ng-show="!$ctrl.showStuffSelectButtons"
-                    class="btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover_bg-gray-200"
+                    class="btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover:bg-gray-200"
                     ng-click="$ctrl.showSelectButtons()"
                 >
                     <i class="fa-solid fa-lg fa-plus-circle text-orange-900 mr-1"></i>
@@ -25,14 +25,14 @@ angular.module("korpApp").component("addBox", {
                 <div class="pos-buttons" ng-class="{'fade-in': $ctrl.showStuffSelectButtons}">
                     <button
                         ng-click="$ctrl.addStructTokenLocal()"
-                        class="block btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover_bg-gray-200"
+                        class="block btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover:bg-gray-200"
                     >
                         <i class="fa-solid fa-arrow-left mr-1"></i>
                         {{ 'before_token' | loc:$root.lang }}
                     </button>
                     <button
                         ng-click="$ctrl.addStructTokenLocal(false)"
-                        class="block btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover_bg-gray-200 mt-2"
+                        class="block btn btn-sm image_button insert_token border-gray-300 transition duration-200 hover:bg-gray-200 mt-2"
                     >
                         <i class="fa-solid fa-arrow-right mr-1"></i>
                         {{ 'after_token' | loc:$root.lang }}